WIP: Simple separation for State objects
There are some tests failing due to missing references in
the propagation code that are not easy to fix.
