+++ src/poliastro/twobody/equinoctial.py
@@ -5,7 +5,10 @@
 import numpy as np
 from astropy import units as u
 
-from poliastro.twobody.core import State
+from poliastro.twobody.base import BaseState
+
+import poliastro.twobody.rv
+import poliastro.twobody.classical
 
 
 def mee2coe(p, f, g, h, k, L):
@@ -30,7 +33,7 @@ def mee2coe(p, f, g, h, k, L):
     return p, ecc, inc, raan, argp, nu
 
 
-class ModifiedEquinoctialState(State):
+class ModifiedEquinoctialState(BaseState):
     def __init__(self, attractor, p, f, g, h, k, L,
                  epoch):
         super(ModifiedEquinoctialState, self).__init__(attractor, epoch)
@@ -75,7 +78,7 @@ class ModifiedEquinoctialState(State):
 
         a = p / (1 - ecc**2)
 
-        return super(ModifiedEquinoctialState, self).from_classical(
+        return poliastro.twobody.classical.ClassicalState(
             self.attractor,
             a * u.km,
             ecc * u.one,
